types: import types in mts from cjs definitions
