+++ package.json
@@ -9,11 +9,10 @@
         "types": "./types/katex.d.ts",
         "default": "./dist/katex.js"
       },
-      "import": 
-        {
-          "types": "./types/katex.d.mts",
-          "default": "./dist/katex.mjs"
-        }
+      "import": {
+        "types": "./types/katex.d.mts",
+        "default": "./dist/katex.mjs"
+      }
     },
     "./contrib/auto-render": {
       "require": "./dist/contrib/auto-render.js",
